+++ sys-auth/app/classes/Exceptions/ValidationFailedException.php
@@ -1,5 +1,10 @@
 <?php
 
+# Prevent direct file access
+if (!defined('APP')) {
+    die(header("HTTP/1.1 403 Forbidden"));
+}
+
 class ValidationFailedException extends Exception
 {
     protected $field;