+++ sys-auth/app/classes/InvalidConfigException.php
@@ -1,6 +1,0 @@
-<?php
-
-class InvalidConfigException extends Exception
-{
-    // ... code
-}