+++ sys-auth/app/classes/PageRenderException.php
@@ -1,6 +1,0 @@
-<?php
-
-class PageRenderException extends Exception
-{
-    // ... code
-}